requirements.txt
python-telegram-bot==21.5
requests
python-dotenv
flask
